fix(runtime): extract plugin migration runner to local module
webpack bundles instrumentation.ts for the Edge build and follows all
dynamic import() calls. Importing @sovereignfs/db directly from
instrumentation.ts causes webpack to bundle better-sqlite3 and pg for
Edge, failing with 'Module not found: Can\'t resolve fs/path'.

The existing pattern avoids this by importing only local module files
from instrumentation.ts (./src/sdk-host, ./src/boot-compat), which pull
in the native-dep packages at a level webpack handles correctly.

Move runIsolatedPluginMigrations() into runtime/src/plugin-migrations.ts
and import it with await import('./src/plugin-migrations') — the same
pattern as sdk-host and boot-compat.
